Add isolation.
This commit adds the isolation for the USERS and
for the TEAMS resources.

Users:
 * Each users will only see the users of its team.
 * Only the super admin or the admin of a team can create
 or delete a user.
 * Only the user itself or an admin of its team can update
 a user.

Teams:
 * Only the super admin can create/delete a team.
 * Only the super admin and the admin of a team can update
 the team.

+++ dci/server/api/v1/teams.py
@@ -28,6 +28,7 @@ from dci.server.common import exceptions
 from dci.server.common import schemas
 from dci.server.common import utils
 from dci.server.db import models_core as models
+from dci.server import dci_config
 
 # associate column names with the corresponding SA Column object
 _T_COLUMNS = v1_utils.get_columns_name_with_objects(models.TEAMS)
@@ -42,8 +43,15 @@ def _verify_existence_and_get_team(t_id):
 
 @api.route('/teams', methods=['POST'])
 @auth2.requires_auth
-def create_teams():
+def create_teams(user_info):
     values = schemas.team.post(flask.request.json)
+
+    user_info_team_id = user_info.get('team_id')
+    # If it's not a super admin, then its not allowed to create a new team
+    if user_info_team_id != dci_config.TEAM_ADMIN_ID:
+        raise exceptions.DCIException("Operation not authorized.",
+                                      status_code=401)
+
     etag = utils.gen_etag()
     values.update({
         'id': utils.gen_uuid(),
@@ -64,11 +72,15 @@ def create_teams():
 
 @api.route('/teams', methods=['GET'])
 @auth2.requires_auth
-def get_all_teams():
+def get_all_teams(user_info):
     args = schemas.args(flask.request.args.to_dict())
 
     query = sqlalchemy.sql.select([models.TEAMS]).\
         limit(args['limit']).offset(args['offset'])
+
+    user_info_team_id = user_info.get('team_id')
+    if user_info_team_id != dci_config.TEAM_ADMIN_ID:
+        query = query.where(models.TEAMS.c.id == user_info_team_id)
 
     query = v1_utils.sort_query(query, args['sort'], _T_COLUMNS)
     query = v1_utils.where_query(query, args['where'], models.TEAMS,
@@ -84,58 +96,83 @@ def get_all_teams():
     return flask.Response(result, 200, content_type='application/json')
 
 
-@api.route('/teams/<ct_id>', methods=['GET'])
+@api.route('/teams/<t_id>', methods=['GET'])
 @auth2.requires_auth
-def get_team_by_id_or_name(ct_id):
-    team = _verify_existence_and_get_team(ct_id)
-    etag = team['etag']
+def get_team_by_id_or_name(user_info, t_id):
+    query = sqlalchemy.sql.select([models.TEAMS]).where(
+        sqlalchemy.sql.or_(models.TEAMS.c.id == t_id,
+                           models.TEAMS.c.name == t_id))
 
-    team = json.dumps({'team': dict(team)}, default=utils.json_encoder)
+    team = flask.g.db_conn.execute(query).fetchone()
+    if team is None:
+        raise exceptions.DCIException("Team '%s' not found." % t_id,
+                                      status_code=404)
+    team = dict(team)
+    user_info_team_id = user_info.get('team_id')
+    if (user_info_team_id != dci_config.TEAM_ADMIN_ID and
+       user_info_team_id != team['id']):
+        raise exceptions.DCIException("Operation not authorized.",
+                                      status_code=401)
+    etag = team['etag']
+    team = json.dumps({'team': team}, default=utils.json_encoder)
     return flask.Response(team, 200, headers={'ETag': etag},
                           content_type='application/json')
 
 
 @api.route('/teams/<team_id>/remotecis', methods=['GET'])
 @auth2.requires_auth
-def get_remotecis_by_team(team_id):
+def get_remotecis_by_team(user_info, team_id):
     team = _verify_existence_and_get_team(team_id)
     return remotecis.get_all_remotecis(team['id'])
 
 
-@api.route('/teams/<ct_id>', methods=['PUT'])
+@api.route('/teams/<t_id>', methods=['PUT'])
 @auth2.requires_auth
-def put_team(ct_id):
+def put_team(user_info, t_id):
     # get If-Match header
     if_match_etag = utils.check_and_get_etag(flask.request.headers)
 
-    data_json = schemas.team.put(flask.request.json)
+    values = schemas.team.put(flask.request.json)
 
-    _verify_existence_and_get_team(ct_id)
+    user_info_team_id = user_info.get('team_id')
+    user_info_role = user_info.get('role')
+    if (user_info_team_id != dci_config.TEAM_ADMIN_ID and
+            (user_info_role != 'admin' or
+             user_info_team_id != t_id)):
+        raise exceptions.DCIException("Operation not authorized.",
+                                      status_code=401)
 
-    data_json['etag'] = utils.gen_etag()
+    _verify_existence_and_get_team(t_id)
+
+    values['etag'] = utils.gen_etag()
     query = models.TEAMS.update().where(
         sqlalchemy.sql.and_(
-            sqlalchemy.sql.or_(models.TEAMS.c.id == ct_id,
-                               models.TEAMS.c.name == ct_id),
-            models.TEAMS.c.etag == if_match_etag)).values(**data_json)
+            sqlalchemy.sql.or_(models.TEAMS.c.id == t_id,
+                               models.TEAMS.c.name == t_id),
+            models.TEAMS.c.etag == if_match_etag)).values(**values)
 
     result = flask.g.db_conn.execute(query)
 
     if result.rowcount == 0:
         raise exceptions.DCIException("Conflict on team '%s' or etag "
-                                      "not matched." % ct_id, status_code=409)
+                                      "not matched." % t_id, status_code=409)
 
-    return flask.Response(None, 204, headers={'ETag': data_json['etag']},
+    return flask.Response(None, 204, headers={'ETag': values['etag']},
                           content_type='application/json')
 
 
 @api.route('/teams/<ct_id>', methods=['DELETE'])
 @auth2.requires_auth
-def delete_team_by_id_or_name(ct_id):
+def delete_team_by_id_or_name(user_info, ct_id):
     # get If-Match header
     if_match_etag = utils.check_and_get_etag(flask.request.headers)
 
     _verify_existence_and_get_team(ct_id)
+
+    user_info_team_id = user_info.get('team_id')
+    if user_info_team_id != dci_config.TEAM_ADMIN_ID:
+        raise exceptions.DCIException("Operation not authorized.",
+                                      status_code=401)
 
     query = models.TEAMS.delete().where(
         sqlalchemy.sql.and_(
